requirements.txt
aiogram>=3.0.0b1
telethon>=1.27.0
tqdm>=4.65.0
PySocks>=1.7.1
PyYAML>=6.0
sentry-sdk>=1.19.1
cryptg>=0.4.0